+++ packages/synapse-interface/components/_Transaction/components/TransactionSupport.tsx
@@ -1,8 +1,11 @@
 import { useTranslations } from 'next-intl'
 import { TRANSACTION_SUPPORT_URL, DISCORD_URL } from '@/constants/urls'
 
-export const TransactionSupport = ({ status }: { status: string }) => {
-  const isReverted = status === 'reverted'
+export const TransactionSupport = ({
+  status,
+}: {
+  status: 'pending' | 'completed' | 'reverted' | 'refunded'
+}) => {
   const t = useTranslations('Time')
 
   return (
@@ -10,11 +13,15 @@ export const TransactionSupport = ({ status }: { status: string }) => {
       id="transaction-support"
       className="flex items-center justify-between w-full py-1 pl-3 pr-1 text-sm"
     >
-      {isReverted ? (
+      {status === 'reverted' && (
         <div>{t('Transaction reverted, funds returned')}</div>
-      ) : (
-        <div>{t("What's taking so long?")}</div>
       )}
+
+      {status === 'refunded' && (
+        <div>{t('Transaction refunded, funds returned')}</div>
+      )}
+
+      {status === 'pending' && <div>{t("What's taking so long?")}</div>}
 
       <div className="flex items-center">
         <a